Moving things over to byte array which should speed things up a bit. Just 36 or so errors left and then I get to compile and figure out how I screwed up the conversion.

+++ Structure.Sketching/Formats/Png/Format/ColorFormats/Interfaces/IColorReader.cs
@@ -28,6 +28,6 @@ namespace Structure.Sketching.Formats.Png.Format.ColorFormats.Interfaces
         /// <param name="pixels">Resulting pixels</param>
         /// <param name="header">Header information</param>
         /// <param name="row">Current row</param>
-        void ReadScanline(byte[] scanline, float[] pixels, Header header, int row);
+        void ReadScanline(byte[] scanline, byte[] pixels, Header header, int row);
     }
 }
